Clean up the tokeniser to handle colons and numeric lines better.
Previously, before the tokeniser was run on each line, the TOK_TEXT
pointer was set to the char following the first colon found in the
line starting at the argv[TOK_CMD]'th char. This breaks when
processing MODE lines which use the extban syntax found in some IRC
servers, as this would kill off the text including and following the
colon in the ban parameter. The beginning of the TOK_TEXT'th command
argument is now handled by detecting a colon immediately following a
delimiter in the tokeniser itself to correctly handle this case.

The tokeniser's special casing for numeric lines (such as 005, 353,
etc) changes the delimiter character mid-tokenisation, which causes
issues when parsing IPv6 addresses in hostmasks (cf. disabling
hostname cloaking on networks where it's supported); further
investigation reveals that various numeric lines are parsed
inconsistently, so punt out of tokenisation on the first delimiter
(i.e. space) after finding a TOK_CMD and a TOK_CHAN whenever a numeric
line is found.

(Speaking of the tokeniser, the fix to correctly detect numeric lines
in a portable manner is now a year old. Happy birthday, commit
257f6c7f34.)